+++ tck/src/main/java/org/apache/jdo/tck/pc/fieldtypes/FieldsOfPrimitiveint.java
@@ -1072,7 +1072,7 @@ false,false,true,false
 
         public String toString() { return this.getClass().getName() + ": "  + identifier;}
 
-        public int hashCode() { return (int)identifier ; }
+        public int hashCode() { return identifier; }
 
         public boolean equals(Object other) {
             if (other != null && (other instanceof Oid)) {
