+++ tck/src/main/java/org/apache/jdo/tck/pc/fieldtypes/MapStringKeyCollections.java
@@ -840,7 +840,7 @@ public class MapStringKeyCollections {
 
         public String toString() { return this.getClass().getName() + ": "  + identifier;}
 
-        public int hashCode() { return (int)identifier ; }
+        public int hashCode() { return identifier; }
 
         public boolean equals(Object other) {
             if (other != null && (other instanceof Oid)) {
